requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
aiosqlite>=0.20.0
anthropic>=0.40.0
httpx>=0.27.0
pymupdf>=1.24.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
playwright>=1.40.0
